Add issue templates
Initially when starting to create the `.github` repository, it was
determined we'd hold off on creating issue templates and such until we
saw recurring patterns. These new templates match the patterns we've
seen so far on the OSCAL-related projects under EasyDynamics and these
are hopefully generic enough to support most projects' needs.