Fix validation errors in GitHub issue forms
For some reason, these did not start appearing on our repositories after
they were added. After playing around with them a bit on a personal
repository, it looks like there are some validation errors. This tries
to resolve those.

This still *may* not work. Issue forms are in beta for public repos and
I wonder if they're not supported from the `.github` repository yet. If
that's the case, I will convert these to issue templates which we may be
able to use.