Add dependency update Starter Workflow
This adds a [starter workflow][1] based on the NPM dependency update
workflow used on the [`EasyDynamics/oscal-react-library`][2] repository.
Adding this as a starter workflow will make it [easy to adopt][3] across
other repositories in the organization. Because each project may need to
customize this workflow (with custom installation steps, NPM versions,
dependency filters, or additional steps), using a starter workflow feels
more appropriate than trying to parameterize all those potential
differences. This will only be shown to repositories that have a file
ending in `package.json`.

[1]: https://docs.github.com/en/actions/using-workflows/creating-starter-workflows-for-your-organization
[2]: https://github.com/EasyDynamics/oscal-react-library
[3]: https://docs.github.com/en/actions/using-workflows/using-starter-workflows